+++ tests/everest/snapshots/test_everest_storage/test_csv_export/config_advanced.yml/batch_df.csv
