+++ tests/everest/snapshots/test_everest_storage/test_csv_export/config_advanced.yml/combined_df.csv
