+++ tests/everest/snapshots/test_everest_storage/test_csv_export/config_minimal.yml/batch_df.csv
